Refactored scythe of oxidation, with some unexpected results #341
My new phrase, "whatever (the item) hit" does not consider weapons
to have touched each other if the global attacker weapon is ranged.
This affects what happens with the chain golem, whose weapon
is ranged, but presumably its "ammunition" remains in contact with the
weapon. Discussion is needed to determine the correct behavior.